metrics: Reorganize code
This commit is a preparation for the further extension of pkg/metrics. There
are no functional changes, only moving code around:
- moved helpers for creating metrics from metricwithpod.go and
  granularmetric.go to per-type files: counter.go, gauge.go and histogram.go
- renamed labels.go to filteredlabels.go
